WORK IN PROGRESS - Continue sequence refactor; get forks to start working
Continued refactoring sequence functionality to ensure robustness.
Importantly, I seem to have gotten branching paths working again
while doing away with clutter in the codebase.

We should be able to remove the 'invert' property from sequences,
instead handling all inversion on the swipe/momentum axis level.
This should dramatically simplify workflows in terms of setting up
branching paths.

Importantly, I've also refactored the fork functionality such that
the JoinTools contains a list of JoinSettings, wherein the previous
and next destinations for each sequence are specified. These destinations
can either be sequences OR forks, and each of the two scenarios
is handled differently.

I also broke out several markers into separate files, as containing them
all in one file was not playing well with Unity's serialization.

Note that lots of cleanup needs to be done, particulary regarding inheritance
schemes and when data is configured. Currently, Unity doesn't support
polymorphic serialization, so much of the configured data needs to be generated
at runtime. I also have not-so-clean workarounds built into the data at the
moment. The upcoming 2019.3 release should have support for polymorphic
serialization though, so I should be able to resolve all of these
issues as a part of a code cleanup soon (the codebase is currently on
Unity 2019.2).

Next steps are to consolidate code, remove old / commented code, and overall
move functionality around so that it's easy to understand and is extensible.
Would also like to test the new axis and fork functionality in a variety of
scenarios to make sure everything works properly. I have a hunch that I'll
need to redo how AxisExtents and ForkExtents relate to each other (right now,
there's no transition between the two states, which makes setting up a fork
more complicated than it should be, will investigate fixing this as I do
the code cleanup.)