Finish first pass at sequence refactor
Resolve last problems related to refactoring sequencing
functionality, including consolidating code inside of the
TouchForkUtils so that it's cleaner and more modular.

Importantly, the TouchForkUtils are structured in such a
way as to search the Joiner's fork data whenever it needs to
get MarkerPlacement information for adjustin ForceForward
and ForceBackward attributes. (Just realized that it would
make more sense to access this information on the AxisMonitor,
so will likely change that.)

Also edited the TouchBranchingPathData so that each branch now
contains a reference to its sequence's own touch data on the
TouchController. Importantly, as I coded this, I realized
that all of a fork's branching paths must contain the
same root config, otherwise we can't generate the data
needed to traverse through forks. There may be a way
around this but it's not a focus for now.

Other than that, renamed some files and did miscellaneous
cleanup.

Note that the OnInitialize method of the JoinMarker is causing
an error when entering PlayMode, so that has been commented out
for now.