Fix bugs related to Lerper; begin refactor of editor tools
Fixed an issue wherein the Lerper wasn't triggering autorun
functionality correctly and also interfering with the autoplayer.
To fix this, a Lerper now uses a threshold modifier when determining
whether it can trigger an autorun sequence or not. In addition,
TriggerInputActionComplete is no longer fired when the
lerper finishes its input; this makes it so that only
additional touch input can trigger modification of a sequence
to resume (the autoplayer was automatically hijacking input
upon the lerper's completion - that's prevented now)

Other than that, began working on the refactor of editor tools.
I'd like to make everything in the editor folder mirror the new
structure of the folders / namespaces. I'm also planning to
abstract the instantiation of different editor tools so that
a Page Builder window can be configured to contain any available
Maestro tools. This will allow us to work in a modular,
package-like fashion while also allowing users to customize
the Page Builder window with the tools they need at any moment.

Next steps are to factor out and consolidate code related to the
Page Builder.