WORK IN PROGRESS - Begin massive overhaul of editor tools
Made substantial changes to the editor tools. Notably, I am in the
process of breaking out all of the editor tools into their
corresponding modules, and creating the architecture needed to
support that.

How this works is that a new Control Panel script, in the root
Editor folder, acts as the brains for the editor tools. Upon
opening a Control Panel, the user is presented with buttons
that correspond to each of these individual modules. Clicking
these buttons, whose UXML names correspond to types inside
of the codebase, will instantiate the class, build the UXML,
and bind the class to the UXML structure.

Importantly, we use Reflection to determine whether the modules
that correspond to the buttons are installed or not. This means
that, once all of this architecture is completed, we should
be able to continue developing the codebase in a package-like
manner, with each module containing its own dependencies,
without the Control Panel needing to know all of the
module-specific functionality.

Certain modules are also dependent on other modules. For example,
the Layout Object module is dependent on the Object Creation module,
which contains methods for renaming items and bool for whether
items should be selected after creation.

I'm still trying to figure out how the dependencies should be
structured for the optimum user experience. As a result, some
functionality is duplicated for the moment (hope to have that cleared
up soon).

Next steps are to continue implementing the rest of the modules,
clean up this functionality, and move onto creating the Timeline Monitor
module.