Add TimelineMonitor functionality; clean up ResponsiveUtils
Finally added a new Timeline Monitor module, which will allow
users to see which timeline clips intersect with the Timeline
playhead, as well as see which clips are currently selected.

I'll likely need to tweak this so that it's actually useful
in real-world scenarios; likely some sort of functionality
tied to the track selection / deselection will eventually be
a part of this feature. That being said, it seems to be good
for now.

Also, I revised the ResponsiveUtils script to be more modular,
as well as renamed a lot of variables within to make it
easier to understand.

Next step for the story engine is to add functionality to add
default Debug and Config tracks. After that, will likely move
onto reworking the scrubbing functionality to work with the new
sequencing refactor, followed by variable reference tools and
refactoring Prepare Scene / Detect Screen Size.