Refactor MomentumApplier to work around axis transitions; make TouchForkExtents edits
After starting work on Violino, I realized that momentum wasn't
being converted between axes correctly and the UX wasn't smooth.
To fix this, I changed how the MomentumApplier works by removing
the reliance on the ConvertMomentum event, which previously
sent a message from an axis switch to the SwipeMonitor to change
the values in its momentum cache.

Now, the MomentumApplier creates its own internal representation of
its momentum force based on the momentum cache whenever a swipe ends.
It then decays its momentum force, the same way the SwipeMonitor does,
but in addition, it contains logic to transfer momentum or get values
from the dominant momentum force when a transition or fork is active.
This makes the functionality much cleaner and removes the dependency
between the sequencing functionality and the swipe monitor functionality.

In addition, I needed to modify how AxisExtents and TouchForkExtents
are configured, because in order for the new MomentumApplier functionality
to work, we need to make sure the transition / fork bool flips on
at the right time.

Also, momentum wasn't working correctly when transitioning from a
a positive to a negative fork - for example, y positive to y negative -
because the isReversing bool would be incorrect (when reversing on a
y positive branch, you are actually moving forward on a y negative branch),
so needed to introduce a check to flip things to the correct state
while going through the fork.

Next step is to move on from the sequence refactor and begin preparing
the Joiner functionality for Chad and Julia's
apocalyptic choose-your-own-adventure story.