POTENTIAL BREAKING CHANGES - Revise modify functionality; various small fixes
IMPORTANT: As part of refactoring the modify functionality, I renamed
various serialized vields, namely aspectRatioBreakpoints and other items
inside of ResponsiveElement, as well as textCollectionBank in the TextLoader.
*THIS MAY BREAK OTHER PROJECTS*. Proceed with caution; I used the
'FormerlySerializedAs' attribute but may need to write import scripts
in order to prevent losing work on Riveted (and perhaps Legend).

Other than that, revised the Modify functionality so that it can
handle multiple layouts and text families simultaneously now. This
should allow for complex interactions with content. For example,
different texts in a single scene may be linked to different layouts
and text collections, which can be activated at any time, and those
texts would sort those dependencies based on priority and execute
them to generate the desired layout.

Also fixed small issues related to the Joiner (adding an 'Active' bool
for Forks), and a bug with AppNavigator wherein duplicate sceneIds were
getting added to its scene dictionary and causing errors.

Next steps are to begin working on Chad and Julia's piece, SHTF, and
make any necessary revision as we move toward a beta version of this
Maestro engine.

the Axis Switches and revise timing with Warren.