Add FloatVariableUtilities, other details for audio manipulation
In order to make audio more interactive, added a new class
that can be used to normalize and otherwise manipulate FloatVariables.
The usage here isn't restricted to audio of course; it can be used
for anything that needs to manipulate float variables in
complex ways.

Other than that, added a custom PitchKey for playing audio and
modified related files accordingly.