Add fixes for Android, including new UseAddressables flag
After testing on Android, it turns out that Android is unable
to load videos from compressed asset bundles, which was causing
errors. The fix for this would be to add videos to separate,
uncompressed asset bundles, then resolve those dependencies
at runtime.

That being said, since we're not planning on using asset bundles in
production anytime soon, I added a temporary fix, a simple
boolean and some code to allow us to not use Addressables if so desired.

This will require that we remove scenes from Addressable groups
every time we want to deploy an Android app for now. Unity is aiming
to fix this issue in 2020, so perhaps by the time we begin using
Addressables in earnest, this will already be resolved.

Also removed code that was modifying the Android sensitivity
during sequence updates (this code is apparently no longer needed).