Add complex event for updating forks; add FrameRateCalculator; other quality-of-life fixes
We neeeded a method for updating forks that is more robust than
manually targeting the joiner, so I created a new complex event,
along with branch keys, called UpdateFork. When assembling a scene,
it's easier to work with ComplexEventPackagers than it is to
directly reference the Joiner when creating interactions, so
should help with workflow. Also added the ability to update a
fork by either branch key OR sequence.

I added a few FrameRateCalculator to help devices running at
slow frame rates (namely Android) keep the correct timing of
our sequences. The reason for this is that, at slower frame
rates, the Update() call doesn't get fired as often inside
of the Autoplayer, which means that sequences will move slower
unless we adjust for that somedhow. FixedUpdate() doesn't help
much, so the FrameRateCalculator solution seems to be the best
way going forward and it seems to work pretty well so far.

Other changes include minor changes to the editor tools, namely
adding the correct layers to newly created objects and sorting
game objects before renaming them.