Add ability to clone object-track placement; add tools for sequencing structures
Two big additions to our editor toolset. A major one
is the ability to clone object-track hierarchies. Given a selection
of game objects, we will find all tracks associated with those
objects, copy the tracks, clone the objects, then map the tracks to
the newly cloned objects in the TrackPlacement clipboard. The tracks
are then ready to be pasted into either the current timeline asset
or a new one.

All of this can be found inside of the TrackPlacement module.

That being said, this relies on our ability to duplicate hierarchies
precisely. So a lot of work went into our DuplicateHierarchy function,
which can be found in Utils. Importantly, I wanted to be able to maintain
prefab connections when duplicating, which required a lot of legwork
to execute correctly. As a bonus, our custom duplicate functionality
is now, as far as I can tell, fully operational. This is exactly
like Unity's Cmd+D duplicate operation, except that it places
duplicated objects adjacent to the original objects.

Other than that, added a toolset for creating sequencing structures,
such as Root Config, Master Sequence, and Sequence Config. These
require a lot of configuration in order to generate correctly,
so it was important to have a tool so as to prevent user error.

Next steps are to remove some unneeded code from the Duplicate
functionality and double check that it works, then add some logic
to determine whether newly created sequence structures should be selected
by default or not.

Note that functionality inside of HierarchySelection that previously
relied on the Duplicate() commands may now be broken.