Refactor playable inheritance; prepare playable migration scripts
Some big changes here. Namely, I refactored the playables so that,
whenever possible, they use the same base clips and behaviours in order
to cut down on code duplication.

For example, previously we had TMProColorClip, SpriteColorClip, and
LerpColorVarClip. I've now refactored the playables so that TMProColorTrack,
SpriteColorTrack, and LerpColorVarTrack now use the same ColorClip instead
of each having their own implementation. This should make the code much
easier to maintain going forward (and substantially cut down on the number
of scripts we have in the codebase).

However, in order to make this work, we need to migrate all of the old clips
in existing projects to the new clips. So I've written a migration
function into LerpToTargetTrack, which contains logic to update our projects
using the new code while (hopefully) not losing any data in the process.

While I was at it, I went ahead and changed the serialized names of all existing
clips to be more generic, and also implemented the use of properties across all
playables to make the code style  more uniform with the rest of the codebase.

Important to note is that all clips that use Vector3s, such as RectTransformPos,
RectTransformRotation, and RectTransformScale, all now use the ResponsiveVector3Clip
and ResponsiveVector3Behaviour functionality, which means that these clips will
have the option of being responsive by default. This also means that
ResponsiveRectTransformPosition and ResponsiveRectTransformScale are deprecated
and will be phased out.

Also important is that I removed duplication between the ResponsiveLerpToTarget functionality
and the LerpToTarget functionality - the former now inherits from the latter.

Next steps are to now carefully migrate Riveted, SHTF, Violino, and Legend of the Girl
and the Condor over to the new implementation. Once that's done, all the obsolete scripts
need to be deleted, and related functionality in the editor tools need to be updated.
Lastly, we'll finally be able to implement the creation of sibling sequences in a relatively
straightforward and easy manner.

After that, I'll be able to continue hardening the editor tools.