Fix migration bug related to double serialization
Realized that the 'ease' field was serialized in both
ResponsiveLerpToTargetBehaviour and LerpToTargetBehaviour,
which was causing the field to not get migrated correctly
when working with Legend.

Aside from that, fixed potential bugs related to determining
cloned asset paths when duplicating timelines.

Also fixed a bug related to the scrubberActive property,
which was returning true incorrectly and firing errors.