Fix serialization bug with RectTransformRelativePos
During the playables refactor, I rewrote RectTransformRelativePos
to inherit from GameObjectReference, and made use of a field
called _childTemplate. However, doing this creates a lot of
duplicate information in the editor, to in order to keep things
clean, I removed the dependency on GameObjectReference.

GameObjectReference as a result is not being used for the moment,
but I will leave in the codebase for potential future
implementations.