Fix RegisterDependencies bug; add PrepareScene to ISkipRegistration; remove ModifySettings dependency
RegisterDependencies was firing errors when encountering null
objects in lists I believe, so added a check to make sure it could
handle that situation.

PrepareScene, likely due to the large amount of responsive elements it houses,
was preventing RegisterDependencies from completing its task, so I gave it the
ISkipRegistration interface for now. I'm planning to move the handling of
responsive elements to a separate script, so this is a temporary fix
until that refactor happens.

ModifySettings is no longer needed, so began removing it from the
codebase, starting with SerializableElement.