Begin consolidating playables; remove duplicate playable functionality
Now that the playables have been refactored, and all projects have
migrated their timeline assets over to the new functionality, that means
we can delete lots of old scripts. For example, TMProColor, SpriteColor,
and LerpColorVar all now utilize the common ColorClip and ColorBehaviour
classes, so we can get ride of TMProColorBehaviour, TMProColorClip,
SpriteColorBehaviour, SpriteColorClip, LerpColorVarBehaviour, and
LerpColorVarClip.

To utilize ColorVar and ColorBehaviour, a particular class needs only
to define a track asset and mixer behaviour to map values onto
any objects in question (i.e. in the case of TMProColor,
we use the TMProColorTrack and TMProColorMixerBehaviour
to map ColorClip and ColorBehaviour functionality onto TMPro objects).

As a result, in an effort to clean up the Playables/ folders, I also
deleted several directories to make scripts easier to find.

Next steps are to continue cleaning up the directories, as well as
continue consolidating editor code given the refactored logic.