Move playables scripts out of child directories
As part of the playables refactor, I decided to move
many scripts to the root of their respective Playables/
folders.

Previously, each group of related playables functionality was
housed in a separate directory. Since we've cut down substantially
on the number of playables scripts in the codebase, and to not veer
too far away from our namespacing conventions (ideally each folder
in the codebase represents a new namespace), I hoisted them
up to allow for better discovery / easier traversal.