Delete old playables functionality; update editor tools and Sibling Sequence functionality
As part of the continuing playables refactor, I finally deleted
the old playabale behaviours and scripts,  and consolidated
functionality within the editor tools to reflect that.

This has reduced a lot of code in the playables functionality,
and makes it easier to navigate the EditAnimationClips module,
which now only has tools for 3 kinds of clips: Color Clip,
FloatClip, and ResponsiveVector3Clip (I still need to add
ResponsiveFloatClip, but that's not priority right now).

Also made minor changes to the AnimationTracks window to
reflect these changes.

I was also able to finally create the SiblingSequence functionality,
which makes use of a new boolean on LerpToTarget tracks that determines
whether tracks should be copied over during sibling creation.

Next steps are to finally move on from this playables refactor and
start cleaning up other parts of the codebase, namely breaking
AppSettings out into DeveloperSettings and UserSettings, and moving
specific settings out of the _AltSalt submodule and into the __Project
submodule.