Add MaestroConfig file to configure project paths; revise AppSettings
Maestro is dependent on certain project paths in order
to render editor tools correctly, as well as dynamically
create folders, settings, and other items. In order to make
this as robust as possible, I introduced a MaestroConfig script,
which will house some of these root paths.

Importantly, this config file is placed in a MaestroConfig directory
outside of the Maestro repo itself - in this way, it becomes part
of the project and is outside of Maestro's version control.

Implementing this required fixing several related utility paths
used throughout the codebase. Everything seems to be working so far.

I also changed AppSettings to no longer utilize persistent variables.
The reason for this is we don't want users to reference variables that
will get recreated with every project - this introduces brittleness
and means that scenes would need to be manually updated whenever moved
between projects.

Next steps then, are to finish ith the AppSettings refactor and
introduce new ways for interfacing with it, or remove some settings
from it so that systems no longer rely on it (namely hasBeenOpened,
musicActive, and soundEffectsActive).

As a bonus, it would be good to introduce relative paths for editor
XML files in keeping with the new MaestroConfig paradigm.