WORK IN PROGRESS - Massive refactor for VariableReference, PrepareScene, IResponsive
As part of the AppSettings refactor, realized that now is the
time to execute some massive refactoring across the codebase
to ensure maximum robustness for projects going forward.

In particular, I've needed to rethink how behaviours hook up
to persistent variables, since it won't be possible to serialize
certain variables between projects. For example, FloatVariables such
as DeviceHeight, SceneAspectRatio, etc., will always change
between projects. As such, we cannot put them into version control
underneath the main codebase, which means they need to be created
dynamically for each project. However, if they get created dynamically,
then how are game objects supposed to hold references to them?

By the same token, let's say we target some of these variables when
setting up scene logic using UnityEvents... In this scenario,
our scene logic would break during migration, because UnityEvents
don't have any way of dynamically searching for equivalent assets.

In order to remedy this, I've started refactoring the VariableReference
scripts in order to store the name of the variables they reference as a
string. Not only that, but I'm making it so the VariableReferences also
require passing in the calling object as a parameter when accessing data -
in this way, we can not only repopulate any missing references
by using the stored name if  need be, we can also store the calling object
and trace which game objects in the scene require the reference,
output that to the console, and make debugging relatively easy.

Refactoring for these multi-project scenarios has other implications as
well, namely that we likely shouldn't be targeting any persistent variables
directly using UnityEvents when setting up game logic. Direct targeting is brittle,
because without a VariableReference as an intermediary,
we can't repopulate the asset if it ever gets lost or we migrate
a scene to a new project. We also can't easily trace which variables
have been changed by which objects in the scene.

For this scenario, I'm planning to introduce a VariableReferenceBehaviour
or something similar, which will allow us to register calling objects whenever
we interact with variables in the editor (this will require modifying
the code for persistent variables and will cause breaking changes in Legend and Violino,
so will need to manually update them or write import scripts).

Next steps then are to continue with this refactor, implementing the refactor
already implemented in  BoolReference across the rest of the VariableReferences
(including the new ScriptableObject reference).

Other than that, I divided PrepareScene into the new scripts PrepareSceneLogic
and DynamicLayoutController. I've also created / revised the related interfaces
IDynamicLayoutElement and IResponsiveBreakpoints to allow RelativeElement
and SyncSequenceToTimeline to become updated on screen size changes based
on priority. (Note: I revised this as part of the AppSettings refactor because I need
to determine how best to dynamically create persistent variables related to layout,
such as DeviceHeight and DeviceWidth, and wanted to have this all clean
as I moved towards finalizing the approach).

All told, AppSettings touches every part of the codebase,
so a large portion ot scripts will continue to be revised before it's done.
Next steps are to continue with the refactor, starting with FloatReference,
then implement VariableReferenceBehaviour, then revise AppSettings to dynamically
create persistent variables related to Sequencing and Touch, then test
to make sure references can repopulate themselves dynamically if the original
data gets destroyed due to migration.