WORK IN PROGRESS - Begin implementing ActionTrigger; revise all persistent variables
Big changes in the works here. The biggest changes are that I'm making it
so that persistent variables MUST be modified through a VariableReference
of some kind; consequently I'm in the process of moving all logic into our new
ActionTrigger and ActionTriggerBehaviour.

The idea here, is that in order to make sure we don't lose our modficationss
of persistent variables when moving scenes between projects, we need to use
variable references as intermediaries. This way, we can maintain a string
with the name of the variable, as well as the target value, to help with migration.
A variable reference, when used, will automatically try to populate the reference
using this stored data; and even if the target variable doesn't exist, a user will
now have all the information they need to recreate the asset.

However, this means that we need to move all logic that targets persistent
variables out of our UnityEvents, which don't allow us to maintain this extra
migration data on their own. That's where the new ActionTrigger comes in.

An ActionTrigger is basically a container that holds a list of different
kinds of actions - BoolActions, FloatActions, etc. All of these actions inherit
from the ActionData base class, and they contain all of the references, operations,
and target values needed to perform particular logic. For example, the BoolAction
contains a BoolReference, an option to Toggle() or SetValue(), and in the case
of SetValue(), a target value of either a bool or BoolVariable.

ActiionTriggers then, allow us to contain lots of disparate functionality in
one place in order to create robust, complex gameplay logic.

Important to note is that I'm in the process of migrating the ConditionResponseTrigger
and (soon) the ComplexEventTriggerPackager over as ActionData so that
we have one source of truth when it comes to implementing all logic.

Since all of our logic will now be contained in one large list within ActionTriggers,
a big part of this commit was also creating the header descriptions necessary for a user
to see, at a high level, what each action is doing without thinking too much (this
required digging into the UnityEvent parameters using SerializedObject and
SerializedProperty).

In addition, since persistent variables will need to be accessed through
variable references anyway, I went ahead and made all persistent variables inherit
from SimpleEvent. The idea here, is that all persistent variables now have the option
of registering listeners, and signaling whenever modifications have occurred.
This means that you can, for example, toggle a BoolVariable on or off using a
BoolAction inside ActionTrigger, then have a separate component listening
to changes in the BoolVariable, then respond accordingly (previously, it
was only possible to listen to regular SimpleEvents and ComplexEvents, not
other kinds of variables).

This last bit of functionality still needs to be finalized and hardened, but
flexibility with regards to setting up logic is where I'm heading.
May lock this down a bit going forward if it turns out to be too messy
when creating logic, but I think it could potentially cut down on bloat
and make things easier to set up overall.

Next steps are to continue implementing ActionData, namely FloatAction and
ComplexEventTriggerAction (potentially with some big quality-of-life changes
to ComplexEventPackager that will cut down on user error in the editor by
prepopulating required and optional parameters).

I also need to actually test everything and remove any gotchas. Lastly, as I
begin wrapping all of this functionality up, will return to AppSettings to ensure
that all common persistent variables can get created dynamically if need be,
with peace of mind knowing that those variables should automatically be repopulated
by variable references through ActionTriggers and ActionData if ever need be.

Last point: It's worth mentioning that this need to dynamically create
variables in AppSettings that led to this entire ActionTrigger refactor
in the first place. Ultimately though I think the codebase will be better for it.