WORK IN PROGRESS - Write migration scripts for UnityEvents; rename ComplexEvent related scripts
As we move towards implementing all logic inside of ActionTriggers, I'd
like to make it so that whenever we call the ActionTriggerBehaviour,
we pass in the calling game object so we can trace the method call.

To do this, I created a new UnityEvent<GameObject> handler, and
proceeded to migrate all of our scripts using standard UnityEvents, namely inside
of SimpleEventListenerBehaviour, Touchable, and ConditionResponseBase.
Legena and Violino both make extensive use of these UnityEvents,
so it's important to move over to the new events without losing any data.

This migration functios, which can likely be used in the future, is now
inside of the MigrationUtils file.

In preparation for implementing complex events inside of the ActionTrigger,
I also renamed some scripts, namely ComplexEventTriggerPackager to simply
ComplexEventPackager, etc. The EventPayload has also been renamed to
ComplexPayload, which matches better with its related scripts.
The renamed ComplexPayloadGenericAction also parallels the
GameObjectGenericAction that is now being used by GenericActionData
via the ActionTrigger.

Next step is to wrap up the GenericAction script, implement ComplexEventPackagers
inside of ActionBehaviour, and add all the other actions.

Note: This commit DOES NOT compile.