Add ActionData for Float, Int, and ComplexEventPackager; fix headings; add SimpleSignals
Wrapped up the ActionTrigger functionality by adding the
last remaining actions, fixing the headings, and resolving a
few bugs related to saving and syncing of the serialized data.

I also created a new base class, called SimpleSignal, that's now
shared between SimpleEvent and ModifiableEditorVariable. All of the
caller and listener registration code previously housed in SimpleEvent
has been moved over to here, and SimpleEventListener and related scripts
have been renamed to SimpleSignalListener, etc.

The idea here is that both ModifiableEditorVariables and SimpleEvents,
while not the same, both can emit signals to any object that's listening.
For ModiableEditorVariables, this occurs whenever the value is changed;
for SimpleEvents, it's whenever it's explicitly triggered via a
SimpleEventTrigger.

Overall, it feels good to know that we can now use SimpleSignalListeners
to respond to changes in persistent variables, and that these changes
are all traceable. That being said, any VariableReferences defined in
code won't make use of SimpleSignal registration capabilities, but
it should make for more intuitive creation of logic inside of the
editor.

It looks to me as though the ActionTrigger functionality is complete.
May run across a few bugs but the mainstay of it seems to be working
properly.

Next step is to move onto creating SimpleSignalReferences or something
similar for SimpleSignalListener, ComplexEventListeneer, references to
layout and text family, and any other scripts relying on persistent
variable references.

After that, planning to move back to AppSettings to begin wrapping
up that refactor as well.