WORK IN PROGRESS - Restructure AppSettings file; begin finalizing access properties
Divided all of the AppSettings into four groups:
SystemSettings, UserData, InputData, and DebugPreferences.

The idea here is that, by having one entry point for all of
these settings, we can dynamically create ones if any are ever
missing. This also means that any new components we add need
hold reference to AppSettings in order to hook into any
existing functionality.

Importantly, I created another layer of separation for
UserData and InputData. The idea here is that, in order
to keep the entire system as robust and modular as possible,
it's possible to create different profiles and groups of
settings.

In the case of UserData, each profile contains preferences for
swipe sensitivity, autoplay, etc. In general, we will only ever have
one user. That being said, if for ever reason we want to introduce
local multiplayer, it'll become as simple as setting up the
input systems (SwipeMonitor, InputControllers) with a new key for
its UserPreferences, which will get created dynamically as soon
as AppSettings attempts to access them.

Regarding InputData, I created a layer of separation so that we
can multiple active groups of inputs going at once. For example,
whenever we start a Tutorial, perhaps we want those swipe settings
to be different than the swipe settings in the main scene. So with
this setup, creating a new batch of settings for the Tutorial becomes
as easy as defining a new custom key, assigning it to the
components in question, and having those elements then get dynamically
populated.

If we ever wanted to do local multiplayer wherein each user
had a separate set of input controllers, swipe monitor, etc.,
this would also be possible with relatively little effort.

Next steps are to continue finalizing / hardening this feature,
then move onto testing.

NOTE: This commit does not compile.