WORK IN PROGRESS - Add new variable references; substantial refactoring
In the process of finalizing the AppSettings refactor, I needed
to modify most of the scripts in the codebase to access the variables.
I also realized that every persistent variable needs its own associated
reference class in order for us to repopulate the field correctly, so
new references have been created.

Importantly, the InputGroupKey and UserDataKey are now being used as keys
inside of InputData and UserData (using CustomKey for I realized is prone to
user error).

As an aside, also began adding more robust functionality for handling of
text color via the newly created TMProColorBehaviour (I started with character
color animation).

While in the last steps of AppSettings refactor, I realized that even when
references are populated dynamically, those changes don't currently get saved
when references exist on prefab instances. So next step is to introduce
functionality that will allow us to find the serialized properties of references
inside of prefab instances so that we can make sure we're saving any
repopulated references if need be.