Add functionality to dynamically repopulate references
Added what I think is the last missing piece of the AppSettings refactor -
dynamically repopulating all references throughout the codebase.

Previously, this repopulation was being done haphazardly on the fly
whenever a variable was accessed through the reference. I was running
into issues with prefab using this method though - the changes to
prefabs weren't being saved after finding the variables.

The reason for this was that we needed to use SerializedObject and
SerializedProperty to apply changes and save them reliably in a scene.

With that in mind, I refactored the ActionTriggerBehaviour, variable
references, condition responses, and other items to ensure that they
can find the location of the serialized variable reference. This involves
passing a string containing the path, from the parent object, all
the way down to the reference in question.

Parent objects in this case should always be MonoBehaviours or some
other UnityObject where serialized data is kept.

Importantly, all of the references inside of AppSettings bypass this feature,
mostly because they already dynamically repopulate themselves using the
RefreshDependencies methods, and making AppSettings work in the same
was as other references would require a refactor of SerializableDictionary
in order to allow us to access the serialized data.

Since this works just fine as is, I decided to leave it, instead introducing
a bool called IsSystemReference on the ReferenceBase to omit the repopulation
functionality.

As far as I can see, all objects, including serialized ones, that use references
seem to be able to repopulate themselves if a persistent variable is ever
deleted and later recreated. This should also be handy for sharing functionality
between projects, allowing us to build reusable systems and components.

Next steps are to do some housekeeping. I'd like to clean up unused code,
clean up the ActionTriggerBehaviour (namely making it so BoolActions, FloatActions,
etc. always use variable references when it comes to target values), and add
list functionality to SimpleEventTriggers.

After that, it'll be some testing to make sure that the new AppSettings
functionalty works properly (will need to do introduce a lot of #if Unity_Editor
conditionals to make sure we're not carting unused authoring data into builds).