Clean up action data items; small revisions to AppSettings to make sure settings save
In addition to the above, I revised the LerpFloatVarMixerBehaviour
and LerpColorVarMixerBehaviour to be compatible with the new
refactor (modifying variables now requires registering the
calling game object).

As a consequence, I was able to update the GradientShift script -
we can make it so that it only updates when its variables are
modified, rather than needing to process refreshing in the
Update() function.