Refactor fader functionality; modify prefabs to work with new app settings system
Made the AppNavigtor work in conjunction with a new Fader script, which
together handle the complex task of loading / unloading scenes so that
our transitions are smooth.

Important to note is that they make use of a lot of complex events, with
many different options and custom keys that can be used to specify the
loading behaviour.

Also combined this with a new ProgressBar script that can be optionally
used to display a loading bar during certain transitions. (This hasn't
been thoroughly tested yet, though in principle seems to be working well).

While testing, I also modified out prefabs to use the new AppSettings variables
by default. This seems to be working well. Also modified a few playable
tracks a bit to accommodate this new functionality.

Next steps are to work on sequence scrubbing, refactor ComplexEventconfigurableTrigger
to utilize variable references instead of straight variables, and move onto
extensive testing and hardening of all these new features.